Fix single precision builds
