+++ src/Equations/viscoelastic2/Kernels/Interface.hpp
@@ -49,7 +49,8 @@ namespace seissol::kernels {
     struct alignas(ALIGNMENT) LocalTmp {
       alignas(ALIGNMENT) real timeIntegratedAne[tensor::Iane::size()]{};
       alignas(ALIGNMENT) std::array<real, tensor::averageNormalDisplacement::size()> nodalAvgDisplacements[4]{};
-      GravitationalFreeSurfaceBc gravitationalFreeSurfaceBc{};
+      GravitationalFreeSurfaceBc gravitationalFreeSurfaceBc;
+      LocalTmp(double gravitationalAcceleration) : gravitationalFreeSurfaceBc(gravitationalAcceleration) {};
     };
     LTSTREE_GENERATE_INTERFACE(LocalData, initializers::LTS, cellInformation, localIntegration, dofs, dofsAne, faceDisplacements)
   LTSTREE_GENERATE_INTERFACE(NeighborData, initializers::LTS, cellInformation, neighboringIntegration, dofs, dofsAne)
