+++ src/Initializer/parameters/SeisSolParameters.h
@@ -1,21 +1,6 @@
 #ifndef SEISSOL_PARAMETERS_H
 #define SEISSOL_PARAMETERS_H
 
-// #include <cstdint>
-// #include <string>
-// #include <array>
-// #include <unordered_set>
-// #include <yaml-cpp/yaml.h>
-//
-// #include <xdmfwriter/XdmfWriter.h>
-//
-// #include "Checkpoint/Backend.h"
-// #include "Initializer/parameters/DRParameters.h"
-// #include "Geometry/CubeGenerator.h"
-// #include "Geometry/MeshReader.h"
-// #include "Initializer/time_stepping/LtsParameters.h"
-// #include "SourceTerm/typedefs.hpp"
-
 #include "ParameterReader.h"
 
 #include "CubeGeneratorParameters.h"
