+++ src/Initializer/parameters/SourceParameters.h
@@ -14,30 +14,7 @@ struct SourceParameters {
   std::string fileName;
 };
 
-inline SourceParameters readSourceParameters(ParameterReader& baseReader) {
-  auto reader = baseReader.readSubNode("sourcetype");
-
-  const auto type = reader.readWithDefaultEnum(
-      "type",
-      PointSourceType::None,
-      {PointSourceType::None, PointSourceType::FsrmSource, PointSourceType::NrfSource});
-  auto readFilename = [&reader](bool enabled) {
-    std::string fileName;
-    if (enabled) {
-      fileName = reader.readOrFail<std::string>("fileName", "No source file specified.");
-    } else {
-      reader.markUnused("fileName");
-    }
-
-    return fileName;
-  };
-  const auto fileName = readFilename(type != PointSourceType::None);
-  reader.warnDeprecated({"rtype", "ndirac", "npulsesource", "nricker"});
-  reader.warnUnknown();
-
-  return SourceParameters{type, fileName};
-}
-
+SourceParameters readSourceParameters(ParameterReader& baseReader);
 } // namespace seissol::initializers::parameters
 
 #endif
